rename opportunity table files to singular to match sibling tables
